Added the unit tests.

+++ bitvector_test.go
@@ -16,12 +16,12 @@ func assert( t *testing.T, value bool, format string, args ...interface{} ) {
 func TestAllocationSize(t *testing.T) {
 	// test single byte allocation
     for i := 1; i <= 8; i++ {	
-		v := New(uint(i))
+		v := New(i)
 		assert(t, len(v.bits)==1, "wrong allocation size for iteration %d", i)
 	}
     // test double byte allocation
     for i := 9; i <= 16; i++ {	
-		v := New(uint(i))
+		v := New(i)
 		assert(t, len(v.bits)==2, "wrong allocation size for iteration %d", i)
 	}
 
@@ -45,18 +45,18 @@ func TestSetGet(t *testing.T) {
     assert(t,v.Get(0),"didn't set correctly")
     assert(t,v.GetInt(0)==1,"didn't set correctly")
     for i := 1; i < 8; i++ {
-		assert(t,!v.Get(uint(i)),"should be 0")
-		assert(t,v.GetInt(uint(i))==0,"GetInt should be 0")
+		assert(t,!v.Get(i),"should be 0")
+		assert(t,v.GetInt(i)==0,"GetInt should be 0")
 	}
 
     // set all bits in w
     for i := 0; i < 16; i++ {
-        w.Set(uint(i),true)
+        w.Set(i,true)
     }
     // and check...
     for i := 0; i < 16; i++ {
-        assert(t,w.Get(uint(i)),"should be 1")
-		assert(t,w.GetInt(uint(i))==1,"GetInt should be 1")
+        assert(t,w.Get(i),"should be 1")
+		assert(t,w.GetInt(i)==1,"GetInt should be 1")
     }
 }
 
@@ -65,20 +65,20 @@ func TestNot(t *testing.T) {
 	w.Not()
     // and check...
     for i := 0; i < 16; i++ {
-        assert(t,w.Get(uint(i)),"should be 1")
-		assert(t,w.GetInt(uint(i))==1,"GetInt should be 1")
+        assert(t,w.Get(i),"should be 1")
+		assert(t,w.GetInt(i)==1,"GetInt should be 1")
     }
 	w.Not()
     // and check...
     for i := 0; i < 16; i++ {
-        assert(t,!w.Get(uint(i)),"should be 0")
-		assert(t,w.GetInt(uint(i))==0,"GetInt should be 0")
+        assert(t,!w.Get(i),"should be 0")
+		assert(t,w.GetInt(i)==0,"GetInt should be 0")
     }
 	w.Not()
     // and check...
     for i := 0; i < 16; i++ {
-        assert(t,w.Get(uint(i)),"should be 1")
-		assert(t,w.GetInt(uint(i))==1,"GetInt should be 1")
+        assert(t,w.Get(i),"should be 1")
+		assert(t,w.GetInt(i)==1,"GetInt should be 1")
     }
 }
 
@@ -88,14 +88,14 @@ func TestNotPackage(t *testing.T) {
     assert(t,v!=w,"v and w are the same")
     // and check...
     for i := 0; i < 16; i++ {
-        assert(t,w.Get(uint(i)),"should be 1")
-		assert(t,w.GetInt(uint(i))==1,"GetInt should be 1")
+        assert(t,w.Get(i),"should be 1")
+		assert(t,w.GetInt(i)==1,"GetInt should be 1")
     }
     z := Not(w)
     // and check...
     for i := 0; i < 16; i++ {
-        assert(t,!z.Get(uint(i)),"should be 0")
-		assert(t,z.GetInt(uint(i))==0,"GetInt should be 0")
+        assert(t,!z.Get(i),"should be 0")
+		assert(t,z.GetInt(i)==0,"GetInt should be 0")
     }
 }
 
@@ -130,7 +130,7 @@ func BenchmarkNotManual(b *testing.B) {
 	v := New(VECTOR_LEN)
     for i := 0; i < b.N; i++ {
 		for j := 0; j < VECTOR_LEN; j++ {
-			v.Set(uint(j),true)
+			v.Set(j,true)
 		}
     }
 }
